+++ src/solver/conv_asm_implicit_gemm_gtc_bwd.cpp
@@ -29,7 +29,7 @@
 #include <miopen/generic_search.hpp>
 #include <miopen/gcn_asm_utils.hpp>
 #include "implicitgemm_util.hpp"
-#include "asm_implicit_gemm.hpp"
+#include <miopen/conv/asm_implicit_gemm.hpp>
 
 namespace miopen {
 namespace solver {
